requirements.txt
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4
numpy==1.26.2
pymongo==4.6.0
python-dotenv==1.0.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
streamlit-authenticator==0.2.3
requests==2.31.0
pytz==2023.3 